Move recorded de-serialization errors to a folder
We move the recorded jsonization and xmlization errors to separate
folders, so that we can easily trace whether the inputs changed
(directories `Json/` and `Xml/`) or the outputs (`JsonizationError/` and
`XmlizationError/`). This separation also allows us to re-record the
errors in an easier way, as we can simply delete the whole directory and
re-run the tests.