Avoid global variables in tests
We should not use global variables in tests whose construction might
throw an exception, as the program immediately exits and no error
message is written to the user.

For example, if the environment variable pointing to the test data is
not set, the global variable specifying a subdirectory in the test data
would simply cause an exception, but the user is oblivious what caused
the problem.