Fix update script so that it builds (#28)
The update script had to be adapted as well as the BigInt library so
that MSVC now builds everything. Additionally, we also have to fix
common XML-ization module in tests.

Previously, a couple of issues in BigInt and common XML-ization
prevented the compilation. Namely, we compared signed and unsigned
values. The update script was too sparse on messages and we could not
figure out what went wrong.