Use a virtual machine for C++ regexes
The C++ standard library comes with a very slow engine for regular
expressions in most implementations. For example, it seems that GNU
C++ compiler runs in exponential time, see this bug report:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=93502

This is impractical for most real-world inputs. Therefore, we implement
a virtual machine for matching regular expressions which runs in ``O(m *
n)`` time (where ``m`` is the number of instructions in the regular
expression and ``n`` is the length of the input).

The original approach has been described in:
Thompson, K., "Regular expression search algorithm", ACM 11(6) 1968

while we follow this blog post:
https://swtch.com/~rsc/regexp/regexp2.html

This corresponds to the change effectuated in the code generation from
[aas-core-codegen a2f329f0].

[aas-core-codegen a2f329f0]: https://github.com/aas-core-works/aas-core-codegen/commit/a2f329f0