Add cross-platform support for tests
In the tests we access environment variables. Until now we used getenv_s, as
this is the preferred way on Windows. (getenv is considered insecure in parallel
environments, however, we don't have such an environment.) However, this is a
Microsoft extension and thus not available to other compilers. We extend the
tests to use the platform appropriate function for access to environment
variables.