Add C++ snippets for code generation
In the past the generation of the code in this repo happened as part of
the release process in `aas-core-codegen`. Updating aas-core3.0-cpp
then merely was a step of copying test data over. However, this causes
deep coupling between these two repos. To uncouple the repos, we move
the snippets to this repo and call `aas-core-codegen` from the
dev_scripts, instead.